requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
rapidfuzz
pillow
numpy
openai
cryptography
PyYAML
python-multipart
aiofiles
